Refactor immutability to trait
