+++ src/Scalars/Immutable/ImmutableArrayOfBoolean.php
@@ -4,10 +4,13 @@ declare(strict_types=1);
 
 namespace ArrayOf\Scalars\Immutable;
 
-use ArrayOf\ImmutableArrayOf;
+use ArrayOf\ArrayOf;
+use ArrayOf\Traits\Immutable;
 
-final class ImmutableArrayOfBoolean extends ImmutableArrayOf
+final class ImmutableArrayOfBoolean extends ArrayOf
 {
+    use Immutable;
+
     protected function typeToEnforce(): string
     {
         return self::SCALAR_BOOLEAN;
