+++ src/Scalars/Immutable/ImmutableArrayOfFloat.php
@@ -4,10 +4,13 @@ declare(strict_types=1);
 
 namespace ArrayOf\Scalars\Immutable;
 
-use ArrayOf\ImmutableArrayOf;
+use ArrayOf\ArrayOf;
+use ArrayOf\Traits\Immutable;
 
-final class ImmutableArrayOfFloat extends ImmutableArrayOf
+final class ImmutableArrayOfFloat extends ArrayOf
 {
+    use Immutable;
+
     protected function typeToEnforce(): string
     {
         return self::SCALAR_DOUBLE;
