+++ src/Scalars/Immutable/ImmutableArrayOfInteger.php
@@ -4,10 +4,13 @@ declare(strict_types=1);
 
 namespace ArrayOf\Scalars\Immutable;
 
-use ArrayOf\ImmutableArrayOf;
+use ArrayOf\ArrayOf;
+use ArrayOf\Traits\Immutable;
 
-final class ImmutableArrayOfInteger extends ImmutableArrayOf
+final class ImmutableArrayOfInteger extends ArrayOf
 {
+    use Immutable;
+
     protected function typeToEnforce(): string
     {
         return self::SCALAR_INTEGER;
