+++ src/Scalars/Immutable/ImmutableArrayOfString.php
@@ -4,10 +4,13 @@ declare(strict_types=1);
 
 namespace ArrayOf\Scalars\Immutable;
 
-use ArrayOf\ImmutableArrayOf;
+use ArrayOf\ArrayOf;
+use ArrayOf\Traits\Immutable;
 
-final class ImmutableArrayOfString extends ImmutableArrayOf
+final class ImmutableArrayOfString extends ArrayOf
 {
+    use Immutable;
+
     protected function typeToEnforce(): string
     {
         return self::SCALAR_STRING;
