+++ tests/Unit/Fixtures/ValidScalarImmutableArrayOf.php
@@ -4,10 +4,13 @@ declare(strict_types=1);
 
 namespace ArrayOfTest\Unit\Fixtures;
 
-use ArrayOf\ImmutableArrayOf;
+use ArrayOf\ArrayOf;
+use ArrayOf\Traits\Immutable;
 
-final class ValidScalarImmutableArrayOf extends ImmutableArrayOf
+final class ValidScalarImmutableArrayOf extends ArrayOf
 {
+    use Immutable;
+
     protected function typeToEnforce(): string
     {
         return 'string';
