+++ tests/Unit/Scalars/Immutable/ImmutableArrayOfBooleanTest.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace ArrayOfTest\Unit\Scalars\Immutable;
 
-use ArrayOf\ImmutableArrayOf;
 use ArrayOf\Scalars\Immutable\ImmutableArrayOfBoolean;
+use ArrayOf\Traits\Immutable;
 use PHPUnit\Framework\TestCase;
 
 final class ImmutableArrayOfBooleanTest extends TestCase
@@ -14,6 +14,6 @@ final class ImmutableArrayOfBooleanTest extends TestCase
     {
         $test = new ImmutableArrayOfBoolean([true]);
         self::assertInstanceOf(ImmutableArrayOfBoolean::class, $test);
-        self::assertInstanceOf(ImmutableArrayOf::class, $test);
+        self::assertContains(Immutable::class, class_uses($test));
     }
 }
