+++ tests/Unit/Scalars/Immutable/ImmutableArrayOfStringTest.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace ArrayOfTest\Unit\Scalars\Immutable;
 
-use ArrayOf\ImmutableArrayOf;
 use ArrayOf\Scalars\Immutable\ImmutableArrayOfString;
+use ArrayOf\Traits\Immutable;
 use PHPUnit\Framework\TestCase;
 
 final class ImmutableArrayOfStringTest extends TestCase
@@ -14,6 +14,6 @@ final class ImmutableArrayOfStringTest extends TestCase
     {
         $test = new ImmutableArrayOfString(['test']);
         self::assertInstanceOf(ImmutableArrayOfString::class, $test);
-        self::assertInstanceOf(ImmutableArrayOf::class, $test);
+        self::assertContains(Immutable::class, class_uses($test));
     }
 }
